+++ src/checkers/twisted.rs
@@ -1,0 +1,207 @@
+//! The Twisted checker.
+//! This module contains the checker used to determine if Twisted is
+//! used by the asset.
+
+use std::collections::HashMap;
+
+use super::HttpChecker;
+use crate::models::reqres::{UrlRequestType, UrlResponse};
+use crate::models::{technology::Technology, Finding};
+use log::{info, trace};
+use regex::Regex;
+
+/// The Twisted checker
+pub struct TwistedChecker<'a> {
+    /// The regexes used to recognize the technology
+    regexes: HashMap<&'a str, Regex>,
+}
+
+impl<'a> TwistedChecker<'a> {
+    /// Creates the checker.
+    /// By doing so, the regex will is compiled once and the checker can be
+    /// reused.
+    pub fn new() -> Self {
+        let mut regexes = HashMap::new();
+	// Example: Twisted/16.5.0 TwistedWeb/8.3.0
+        let header_regex =
+            Regex::new(r"(?P<wholematch>.*Twisted\/(?P<version>\d+\.\d+\.\d+).*)").unwrap();
+
+        regexes.insert("http-header", header_regex);
+        Self { regexes: regexes }
+    }
+
+    /// Check in the HTTP headers.
+    fn check_http_headers(&self, url_response: &UrlResponse) -> Option<Finding> {
+        trace!(
+            "Running TwistedChecker::check_http_headers() on {}",
+            url_response.url
+        );
+        // Check the HTTP headers of each UrlResponse
+        let headers_to_check =
+            url_response.get_headers(&vec!["Server".to_string()]);
+
+        // Check in the headers to check that were present in this UrlResponse
+        for (header_name, header_value) in headers_to_check {
+            trace!("Checking header: {} / {}", header_name, header_value);
+            let caps_result = self
+                .regexes
+                .get("http-header")
+                .expect("Regex \"http-header\" not found.")
+                .captures(&header_value);
+
+            // The regex matches
+            if caps_result.is_some() {
+                info!("Regex Twisted/http-header matches");
+                let caps = caps_result.unwrap();
+                return Some(
+                    self.extract_finding_from_captures(
+                        caps,
+                        url_response,
+                        45,
+                        45,
+                        "Twisted",
+                        &format!("$techno_name$$techno_version$ has been identified using the HTTP header \"{}: $evidence$\" returned at the following URL: $url_of_finding$", header_name)
+                    )
+                );
+            }
+        }
+        None
+    }
+}
+
+impl<'a> HttpChecker for TwistedChecker<'a> {
+    /// Check if the asset is running Twisted.
+    /// It looks in the following HTTP headers:
+    /// - Server
+    /// - X-Powered-By
+    /// and in the "not found" page content
+    ///
+    /// Returns only one finding, otherwise findings would be duplicated each
+    /// time it's found.
+    fn check_http(&self, data: &[UrlResponse]) -> Vec<Finding> {
+        trace!("Running TwistedChecker::check_http()");
+
+        for url_response in data {
+            // JavaScript files could be hosted on a different server
+            // Don't check the JavaScript files to avoid false positive,
+            // Check only the "main" requests.
+            if url_response.request_type != UrlRequestType::Default {
+                continue;
+            }
+
+            let header_finding = self.check_http_headers(url_response);
+            if header_finding.is_some() {
+                return vec![header_finding.unwrap()];
+            }
+        }
+        return Vec::new();
+    }
+
+    /// The technology supported by the checker.
+    fn get_technology(&self) -> Technology {
+        Technology::Twisted
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::checkers::check_finding_fields;
+
+    #[test]
+    fn header_matches() {
+        let checker = TwistedChecker::new();
+        let mut headers1 = HashMap::new();
+        headers1.insert("Accept".to_string(), "text/html".to_string());
+        headers1.insert("Server".to_string(), "Twisted/8.2.1".to_string());
+        let url1 = "https://www.example.com/that?abc=def";
+        let mut url_response_valid =
+            UrlResponse::new(url1, headers1, "the body", UrlRequestType::Default, 200);
+        let finding = checker.check_http_headers(&url_response_valid);
+        assert!(finding.is_some());
+        check_finding_fields(
+            &finding.unwrap(),
+            "Twisted/8.2.1",
+            "Twisted",
+            Some("8.2.1"),
+            Some(url1),
+        );
+
+        let mut headers2 = HashMap::new();
+        headers2.insert("Accept".to_string(), "text/html".to_string());
+        headers2.insert("Server".to_string(), "nginx/1.22.0 (CentOS)".to_string());
+        headers2.insert("X-powered-by".to_string(), "Twisted/7.4.9".to_string());
+        url_response_valid.headers = headers2;
+        let finding = checker.check_http_headers(&url_response_valid);
+        assert!(finding.is_some());
+        check_finding_fields(&finding.unwrap(), "Twisted/7.4.9", "Twisted", Some("7.4.9"), Some(url1));
+    }
+
+    #[test]
+    fn header_doesnt_match() {
+        let checker = TwistedChecker::new();
+        let mut headers1 = HashMap::new();
+        headers1.insert("Accept".to_string(), "text/html".to_string());
+        headers1.insert("Server".to_string(), "Apache/2.4.51".to_string());
+        let mut url_response_invalid = UrlResponse::new(
+            "https://www.example.com/that.php?abc=def",
+            headers1,
+            "the body",
+            UrlRequestType::JavaScript,
+            200,
+        );
+        let finding = checker.check_http_headers(&url_response_invalid);
+        assert!(finding.is_none());
+
+        let mut headers2 = HashMap::new();
+        headers2.insert("Accept".to_string(), "text/html".to_string());
+        url_response_invalid.headers = headers2;
+        url_response_invalid.request_type = UrlRequestType::Default;
+        let finding = checker.check_http_headers(&url_response_invalid);
+        assert!(finding.is_none());
+    }
+
+    #[test]
+    fn finds_match_in_url_responses() {
+        let checker = TwistedChecker::new();
+        let mut headers1 = HashMap::new();
+        headers1.insert("Accept".to_string(), "text/html".to_string());
+        headers1.insert(
+            "Server".to_string(),
+            "nginx/1.22.2 OpenSSL/1.0.2k Twisted/8.1.1".to_string(),
+        );
+        let url1 = "https://www.example.com/test";
+        let url_response_valid =
+            UrlResponse::new(url1, headers1, "the body", UrlRequestType::Default, 200);
+        let url_response_invalid = UrlResponse::new(
+            "https://www.example.com/invalid/path.php",
+            HashMap::new(),
+            "nothing to find in body",
+            UrlRequestType::Default,
+            200,
+        );
+        let findings = checker.check_http(&[url_response_valid, url_response_invalid]);
+        assert_eq!(1, findings.len());
+        check_finding_fields(&findings[0], "Twisted/8.1.1", "Twisted", Some("8.1.1"), Some(url2));
+    }
+
+    #[test]
+    fn doesnt_find_match_in_url_responses() {
+        let checker = TwistedChecker::new();
+        let mut headers1 = HashMap::new();
+        headers1.insert("Accept".to_string(), "text/html".to_string());
+        headers1.insert("Server".to_string(), "Twisted/7.1.4 Apache/2.4.52".to_string());
+        let url_response_invalid1 = UrlResponse::new(
+            "https://www.example.com/abc-1/de-f1",
+            headers1,
+            "the body",
+            UrlRequestType::JavaScript,
+            200,
+        );
+        let findings = checker.check_http(&[url_response_invalid1]);
+        assert!(
+            findings.is_empty(),
+            "Twisted must not be detected on JavaScript URLs to avoid false positive"
+        );
+    }
+}